EWM-241. Bookmark modal
